+++ R/tidymedia-package.R
@@ -4,11 +4,6 @@
 NULL
 
 ## usethis namespace: start
-#' @importFrom assertthat assert_that
-## usethis namespace: end
-NULL
-
-## usethis namespace: start
 #' @importFrom glue glue
 ## usethis namespace: end
 NULL